Update patch set 2
Patch Set 2:

Is okey you to make it opt-in, but taking a change from other user which already sent to gerrit and doing the same he does another way with him commit message should have this gerrit change ownership.

He discovered a way to fix FB crash, so now give proper authorship, or in the best case, push a new patchset on him commit. This is about fixing errors, not who fix errors.

Patch-set: 2