Update patch set 1
Patch Set 1:

I noticed that the new shim mechanism fails at executing the constructor of shim libraries. I'm not sure this is related to the issues we have (the ones Luca mentioned) nor if there are actually constructors that we need to be executed, but shows that the mechanism is not prefect.

Patch-set: 1