Update patch set 1
Patch Set 1:

> What is a constructor of a shim library?

I'm referring to functions declared with "__attribute__((constructor))". Shouldn't these be executed?

Let's say I write a library X with such a function and instruct the linker to load X when library Y is loaded. If I do so, the library declared as constructor is not executed, but it is if I LD_PRELOAD it.

> The commit message here explains one of the reasons that you absolutely want to use the new shim libraries...

Definitely and I really like that approach, it just wasn't working (yet) and this is sort of a temp solution.

Patch-set: 1