Update patch set 1
Patch Set 1:

Hmm, I didn't even notice that Ethan was using a constructor / destructor annotation in that.  I'm 99% sure the constructor never gets called in it and either way it doesn't do anything so that's definitely not your problem.

I'm trying this:

http://review.cyanogenmod.org/#/c/119863/

(and the previous commit, which imports and then makes the 12.1 version of SensorManager work in M).  This seems to be working better for me for my device on cm-13.  My camera still doesn't work, but it gets past some obvious memory corruption problems for me.

Patch-set: 1