Update patch set 1
Patch Set 1:

I actually had a mistake in my test program.  The shim libs are not loading the constructors but android does correctly load them.  That's a bug but not the real problem when we are using "the" shims for SensorManager.  The problem is that causes memory corruption and we need to write different shim libs.

See the commit message for

http://review.cyanogenmod.org/#/c/119863/

Patch-set: 1